move loadFilenames function from builtin-tasks/deploy.ts to lib/files.ts
