+++ packages/algob/test/lib/files.ts
@@ -1,8 +1,10 @@
 import { ERRORS } from "@algo-builder/web";
 import { assert } from "chai";
 
-import { assertDirChildren, assertDirectDirChildren } from "../../src/lib/files";
+import { TASK_TEST } from "../../build/builtin-tasks/task-names";
+import { assertDirChildren, assertDirectDirChildren, loadFilenames } from "../../src/lib/files";
 import { expectBuilderError } from "../helpers/errors";
+import { useCleanFixtureProject } from "../helpers/project";
 
 describe("assertDirChildren", () => {
   it("Should pass for children inputs", async () => {
@@ -43,5 +45,21 @@ describe("assertDirectDirChildren", () => {
       () => assertDirectDirChildren("a", ["a/../b/1"]),
       ERRORS.BUILTIN_TASKS.DEPLOY_SCRIPT_NON_DIRECT_CHILD,
       "b/1");
+  });
+
+  describe("loadFilenames", () => {
+    useCleanFixtureProject("typescript-project");
+    it("Should load ts and js files from test folder", function () {
+      const ls = loadFilenames("test");
+      const expected = ['test/js-test.js', 'test/ts-test.ts'];
+      assert.deepEqual(ls, expected);
+    });
+
+    it("Should throw error if dir name is not \"test\"", function () {
+      expectBuilderError(
+        () => loadFilenames("tests", TASK_TEST), // as dir should be "test"
+        ERRORS.BUILTIN_TASKS.TESTS_DIRECTORY_NOT_FOUND
+      );
+    });
   });
 });
